ai-trust/agent-ops/assets/langgraph-agents/requirements.txt
# Python Version Requirement: 3.9-3.12 (3.13+ not yet fully supported by all dependencies)

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.1.0,<2.2.0

# IBM watsonx dependencies
ibm-watsonx-ai>=1.3.0
ibm-watsonx-gov==1.2.2
ibm-cloud-sdk-core>=3.10.0

# IBM agent analytics - REQUIRED for evaluation metrics to work
ibm-agent-analytics>=0.5.0
ibm-agent-analytics-common>=0.1.0
ibm-agent-analytics-core>=0.9.0

# Missing dependencies from ibm-watsonx-gov (not properly declared)
# These are required but not listed in ibm-watsonx-gov's dependencies
unitxt>=1.0.0
scikit-learn>=1.0.0
textstat>=0.7.0
nltk>=3.8.0
jsonpath-ng>=1.5.3

# LangChain core
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langgraph>=0.3.0,<0.4.0

# LangChain integrations
langchain-ibm>=0.3.0,<0.4.0

# Vector store (upgraded to avoid Rust bindings bug in 1.1.1)
chromadb>=0.4.22

# Document loading
pypdf>=3.17.0

# Web search (optional - for AdvancedRAGEvaluator)
tavily-python>=0.3.0

# Jupyter (for notebook usage)
jupyter>=1.0.0
ipykernel>=6.29.0
ipython>=8.12.0
nbformat>=5.0.0

# Development dependencies (optional)
pytest>=8.0.0
